feat: move git func a single domain
